Create seller listing
Create page for adding a new listing with
the following features:
- item name
- upload images (only 1 atm)
- price
- type/category
- description
- quantity

Apprently in PHP you have to upload/store
the file the user gives you locally before
you can have access to the actual binary,
so new seller photos will go into the uploads/
folder (to be included).

(todo) this is most likely not the best design as
there could be a potential for images with
the same name to either overwrite/not be
accepted by the server.

(todo) various minor bugs and formatting issues